add tna_search_spec
