+++ spec/report/duplicate_places_spec.rb
@@ -3,7 +3,7 @@ require 'tnw_common'
 require 'tnw_common/report/duplicate_places'
 
 describe TnwCommon::Report do
-  dp = TnwCommon::Report::DuplicatePlaces.new('http://localhost:8985/solr/archbishops')
+  dp = TnwCommon::Report::DuplicatePlaces.new('http://localhost:8983/solr/archbishops')
   it 'has a report method' do
     duplicates = dp.report()
     expect(duplicates.keys.length).to be >= 0
